chore: move away from k8s javascript client fork (#3967)
* chore: move away from k8s javascript client fork

The changes in the fork have been released. We can move away from the
fork.

This enables NPM migration.

* ci: make sure to use fresh cache when yarn.lock changes

This is to avoid running into caching issues

* fix: pin down version to 0.18.0

We had to pin the version to 0.18.0 because with 0.18.1 we ran into a problem due to kubernetes client's migration from execa to child_process.spawnSync, see also https://github.com/kubernetes-client/javascript/issues/1013