improvement(logger): various fixes to make basic logger render nicer
In a handful of places the rendered output assumes the fancy logger is
being used and it really doesn't play nice with the basic logger.

This commit adds a logger "type" field to the root Logger and in a
handful of places tweaks a given log message depending on the that type.

This is a prerequisite to making the basic logger the default but we
still need to add some sort of a status indicator at the bottom before
doing that since otherwise it's hard to tell whether Garden is still
executing or waiting for changes when in persistent mode.