refactor(logger): cleanup original implementation + remove fields
The first iteration on the new append-only logger was a bit rushed.

This commit cleans up some to dos, adds code comments, tests, and
more.

We also remove some fields that were previously configurable to ensure
logs are more consistent. In particular we remove the 'section' and
'symbol' fields for some levels and instead have the log/renderers handle
the styling.

E.g., now you can just do `log.warn("oh noes)` as opposed to
`log.warn({ msg: chalk.yellow("oh noes"), symbol: "warning})`.

If we realise we need more control we can always re-introduce it,
but it feels right to start conservatively.

There's a chance this may break the rendering of some log lines but that
should be easy to fix in follow up commits.