feat(dev): support persistent commands (logs, deploy --sync/local etc) in dev
This needed a major refactor, but ultimately results in a nice structure that
we can build rapidly on and add various commands and monitoring features.

The way this is implemented is via a new concept called a `Monitor`. For
example, `SyncMonitor` and `LogsMonitor`. So when you run
`garden logs --follow`, a `LogsMonitor` instance is created and attached for
each Deploy action.

Any number of these can be added to the Garden instance via the `MonitorManager`
instance on `garden.monitors`. When those are found to be active at the end of a
Command execution, the process will wait until the monitors exit.

In the dev command, a single Garden instance is shared across many command runs,
so the monitors are persisted. There's also a new `hide` command (only in the
dev command) to halt any started monitors.