docs(k8s): update terraform-gke example for Bonsai  (#4000)
* Generate kubeconfig for GKE cluster

In order to "one-shot" deploy the example a GKE cluster and our demo
services, we need a kubeconfig. Because of
https://github.com/garden-io/garden/issues/3475, a kubeconfig managed by
Garden should be generated with elevated privileges.

* Set hardcoded values to variables

To improve comprehension and expose the values a user should set for a
project at the top of the file, we set the hardcoded values of the
Garden project file to variables.

* Convert 0.12 modules to 0.13

* Specify Terraform one-shot deployments in docs

The README previously assumed a cluster ready to work with. We change
the instructions to assume no pre-existing cluster.

* Simplify and atomize terraform

Previously the GKE example was an all-in-one terraform file. I've chosen
to make use of battle-tested GKE modules and split the terraform for
easy comprehension.