docs: 0.13 update "Using the CLI" guide (#4108)
* remove outdated section about `garden run service` sub-command

Command `garden run` do not have any sub-commands in 0.13.

* docs: remove `cleanup-cluster-registry` command section

The command is no longer used and has no effect in 0.13.
In-cluster registry support was removed in 0.13. In-cluster builds are still possible with either `kaniko` or `buildkit`.

* docs: use action-based terminology

Added TODO for "Creating an action". There is no command to create any actions yet.

* docs: update `dev` command description

The command is no longer used and has no effect in 0.13.
In-cluster registry support was removed in 0.13. In-cluster builds are still possible with either `kaniko` or `buildkit`.

* docs: remove local dashboard section

The removed section was obsolete. The local dashboard was removed in 0.13, and the semantics of the `dev` command was changed.

---------

Co-authored-by: Orzelius <33936483+Orzelius@users.noreply.github.com>