revert: remove noProject flag from custom commands
Refs: 96f6721 and #3983

In PR #3983 we set noProject=true on the custom commands class.

This was to prevent the entire init logic from re-running for the inner
command when set.

Re-running the init logic means the user is authenticated against Cloud
twice, two buffered event streams are created, the command is tracked
twice, etc etc.

However, setting noProject=true on the outer command means that
variables aren't resolved which means you can't reference them in
custom commands which of course also causes issues.

So despite its flaws, we're reverting to the previous version. But in
general this behaviour hasn't been thought fully through, and the
implementation we're reverting to doesn't play well with Cloud.

In my opinion, we should create a single session for a single custom
command and have primitives that allow for the inner command to "belong" to
the outer custom command and display things appropriately in Cloud.

That's a larger task though and needs to be addressed separately.