fix(k8s): stop retrying in LogFollower when Pod is in terminal phase (#4131)
If the Pod phase is "Succeeded" or "Failed", then Kuebrnetes will not
retry the Pod anymore – it is terminal.

In this case, LogFollower should not try fetching more logs, because
there are no more logs to fetch.

This fixes the issue where we wait for 20 seconds and log "Failed to
finish streaming logs: Timed out after 20 seconds. Some logs might be
missing in the verbose log output, or in Garden Cloud." in case of
failed tests.

This also fixes missing Pod detection (another case where we do not want
to retry).

This also reduces the time it takes to run the test suite.