fix(monitors): ensure status isn't overwritten due to a race condition
Previously, the monitor statuses known to the monitor manager could be
wrong due to a race condition when toggling monitors.

Consider the following:

- T1: Log monitor A is requested via websocket
- T2: Log monitor A is aborted via websocket
- T3: The stop function from T2 returns and sets the status to stopped
- T4: The start function from T1 returns and set the status to starteed

In this case the final status is incorrectly 'started' as opposed to
'stopped'.

It's not clear if this actually causes issues but in any case it's
confusing and makes it hard to debug other issues around monitors if we
can't trust the statuses displayed by the manager.