fix(server): ensure fresh command instance per request
Previously we were re-using the same command instance for every server
request (created at start time via getCoreCommands()).

This meant that multiple requests would all subscribe to the same
instance.

Concretely, if request A, asks for logs for service X and request B,
asks for logs for service Y, they would both subscribe to the same logs
command.

Now, when service X writes logs, the callbacks for both request A and
request B would be called and to the end user it looks like both
services X and service Y are writing the logs.