improvement(monitors): allow multiple commands to subscribe to a monitor
This is a follow up to: ac54b93ef8680f12e86b24b52dcf5221af5e8be7

The previous commit ensured that different ws requests would not
subscribe to the same command instance.

The introduced an issue around monitors: Now we have multiple command
instances but they weren't being set on properly on a given monitor.

The commit fixes that by treating commands as subscribers to monitors
and allowing for multiple command to subscribe to the same monitor at
the same time.