fix(serve): ensure buffered event stream works with serve commands
When starting the CLI, a buffered event stream is created that subscribes
to events from a Garden instance.

However, when running a serve command, a new Garden instance is created
via the "reload" function which the buffered event stream wasn't
subscribing to.

This commit fixes that by setting the new instance explicitly on the
buffered event stream.

This is pretty hacky and just a workaround to unblock things which will
be removed when we introduce a proper instance manager.