fix(server): ensure server logs aren't emitted twice
Before this, server logs would be emitted twice over ws. Once via the
'commandOutput' event which a request would subscribe to and once as a
'logEntry' event.

By setting the logger instance to the VoidLogger for the logs when
executed by the server, we eliminate the latter and only emit the log
once.

Note that this is a bit hacky and means that other log entries that the
logs command creates are also ignored.