fix(k8s): fix namespace caching logic for dev cmd
Because the `delete namespace` command can be called interactively
(which deletes the default app namespace), and because of the `reload`
command, we needed a different approach to caching namespace statuses in
the Kubernetes plugin.

The new approach scopes the cache by provider UID, and clears the cache
for a given provider UID if the `delete namespace` command is called.

We also introduced an async lock to the `ensureNamespace` helper, since
the first `deploy` command after a `delete namespace` command can result
in more than one concurrent `ensureNamespace` call for the same
namespace.

Fixes https://github.com/garden-io/garden/issues/4223.