improvement(dev): handle config changes
If one or more Garden configs is changed during a `dev` command, we
lazily reload the project when the next command is run.

This approach avoids multiple reloads when the user is modifying their
Garden configs.

The old Watcher class was reduced down to its bare essentials for this
purpose (up to this point, it had been unused in 0.13).