feat: add "sync status" command
This commit adds a command for getting the current status of all
configured syncs.

In this first iteration the command is not persistent but we may decide
to add a "--follow" or "--monitor" flag later.

The main reason I'm not doing it here is because even if we're
monitoring the sync statuses, we don't get the "sync stop" events from
Mutagen which is one of the main events we'd want to follow in the first
place.

So since sync monitoring would be incomplete and you'd need to re-run
the command to get the latest status I decided leave it out for now.