previously the error was
"Namespace garden-system doesn't exist and Garden was unable to create
it. You may need to create it manually or ask an administrator to
do so."

which is plain misleading.

I tried to implement logic that waits for the termination to complete,
but the `ensureNamespace` function is called multiple times which makes
it hard and I judged it not worth the effort.