chore: update terraform-gke example configurations
The previous config used a `template = file()` call, which implicitly
uses the hashicorp/template provider, which has been deprecated. With
its last release being made 3 years ago, there are no (and will never
be) builds for arm64 based Macs. This meant that the example was not
runnable on Apple Silicon machines, as the providers could not be
loaded. Newer terraform versions have a builtin `templatefile` function
which provides similar functionality out of the box.