fix(server): do not filter subscription when emitting events
NOTE: This is just to get the Core<>Cloud integration working and
ensure events are emitted.

I suspect we don't need to check the subscription key for log entries
since those should probably just be streamed by default, assuming a
connection is live, but not sure what we want to do with events.

So this is just to unblock myself but probably needs revisting.