chore(terraform-plugin): use Deploy action terms instead of Modules
* docs: update terraform-plugin docs for 0.13

* chore: terraform plugin: use Deploy action terms instead of Modules

BREAKING CHANGE: `garden plugins terraform apply-module` is now `garden
plugins terraform apply-deploy`, and similar change for the `plan-*` and
`destroy-*` commands.

* chore: update terraform-gke example configurations

The previous config used a `template = file()` call, which implicitly
uses the hashicorp/template provider, which has been deprecated. With
its last release being made 3 years ago, there are no (and will never
be) builds for arm64 based Macs. This meant that the example was not
runnable on Apple Silicon machines, as the providers could not be
loaded. Newer terraform versions have a builtin `templatefile` function
which provides similar functionality out of the box.

* docs: more terraform-plugin terminology updates

* chore: terraform-plugin: use new command names in tests

* chore: terraform-plugin docs wording improvement

Co-authored-by: Orzelius <33936483+Orzelius@users.noreply.github.com>

* chore: terraform-plugin docs wording improvement

Co-authored-by: Orzelius <33936483+Orzelius@users.noreply.github.com>

* chore: terraform-plugin docs: align file naming with pulumi plugin

* chore: terraform-plugin: use -action suffix in command names

---------

Co-authored-by: Orzelius <33936483+Orzelius@users.noreply.github.com>