feat(core): added sync restart command
This command stops and starts any active syncs for the relevant
Deploy actions.

This is useful for recovering from error states for syncing tools.

Also added diagnostic message logging for failure states to the
`sync status` command.

For Mutagen-based syncs, we provide clear information to the user why
the sync has been halted, and suggest running the `garden sync reset`
command after making sure that the relevant directories are in the
expected state.