improvement(commands): add --skip-detail and --only-deploys flags
This commit adds a couple of "internal" flags to the sync status and get
status commands that make it easier for Cloud and Desktop to consume
their output.

Note that the "garden get status" command will be replaced with a top
level "garden status" command and that these changes aren't really meant
to be user facing. Rather just to unblock a few things on the Cloud
side.