docs(0.13):  quickstart revamp (#4283)
* `garden dev` as default

Switch to `garden dev` as default and explain its status as REPL.

* Remove "plugin" preamble

We mention a `local-kubernetes` plugin almost immediately then describe a pluggable architecture. Such details are best left outside the quickstart which is trying to get a user going as quickly as possible without confusion.

* Stand off code blocks with spaces

* Clean macOS tab title

All macOS devs are likely to be using Homebrew already

* Assume Docker Desktop for local Kubernetes

Docker Desktop supports macOS, Windows and Linux. This guide assumes you
have Docker Desktop installed and running on your local machine. Other
Kubernetes environments are supported, but not covered in this guide.
This makes for a tighter, cleaner quickstart with one happy path.

* Clone and change directory in one command

* Revamp intro to Garden

For users who link in directly to the quickstart, they may want to know
the Cliff's Notes version of what Garden is and why they should use it.
We define Garden as a single pane of glass for users.

* Clean up outro of quickstart guide

* Add deploy step and instructions for running tests

* Update quickstart.md

* Add deploy gif

* Change percentage word

* Add test gifs

* Move deploy to its own line

* Use branded Discord invite link

* Cleanup command center grammar

* Add visual quickstart info

* docs(quickstart): add protocol to discord link url

---------

Co-authored-by: Jhan Silva <jhancarlossilva.c@gmail.com>
Co-authored-by: Emanuele Libralato <10ko@users.noreply.github.com>