fix(core): reintroduce kind-specific cmd outputs
In 0.12, we included keys for build, deploys and tasks in the command
result for commnds that used the task graph (e.g. the build & deploy
commands).

In 0.13, we initially removed these in favour of the `graphResults`
object. Here, we reintroduce them (with most fields still there). This
should prevent breakages e.g. in workflows where referencing command
outputs is relatively common.

In a future PR, we should get rid of the `graphResults` key entirely
from command results. Doing so will require reworking some of our unit
tests (since some suites rely on the dependency structure of the results
to verify which tasks were actually run).