fix: log errors
if just the error was passed to the log entry it was not logged
previously. This might have swallowed up quite a few errors in the past.

Most of the time the error is formated into the `msg` on the log entry
so I added logic that checks whether the entry already includes the
the error and if so doesn't duplicate it.