fix(k8s): correctly resolve manifests when `build` is set
This ended up being a bit of a refactor, since we already had a TODO in code
around resolving manifests when using build dependencies, and that needed to
be resolved to properly address the issue.

`kubernetes` module/Deploy statuses now rely on a metadata ConfigMap, in
principle akin to how Helm chart statuses are resolved. This avoids having
to actually perform builds before working out the manifests that belong to
a Deploy action.

Along the way, avoided superfluous API calls when comparing resources and
resolving their statuses.