refactor: improved plugin definition SDK interface, phase 1
This introduces a more friendly wrapper for defining plugins. One of the key
benefits is that plugin authors don't need to think as much about types
and implicitly get solid type-safety when defining handlers.

To start, the `hadolint` plugin has been ported to the new SDK.

This will need more iteration, particularly relating to ease of use when
defining a plugin and its handlers across multiple modules. A good next step
will be to convert the `exec` plugin. Once this is more fleshed out we'll
also want to document this more thoroughly.

The implementation involves using Zod schemas instead of the non-typesafe Joi
schemas. The underlying framework is unchanged in those terms, and the Zod
schemas are wrapped with a custom Joi method. This allows gradual migration.