fix: ensure build context on ´ready´ status builds
Ensure build context being available even if the plugin handler returns
`ready` status for the action. For example if a container is already
built but the build context has additional files needed for the Deploy.

Did some slight optimization by first checking if the context has been
created previously or the directory already exists. If it does we assume
it to be up to date.