improvement(events): improve types/consistency for action status events
This commit makes several improvements / fixes to the action status events
sent to Cloud.

A) The types for action status events we emit to Cloud are
narrower. We e.g. ensure that the "completedAt" field is required with
the relevant states, that action state and detailed status are
consistent, add missing fields, and more.

B) Ensure the event emissions are handled in a consistent manner by
introducing a decorator that we call at the task level, not the router
level as before (see also C) below).

C) Ensure superfluous events are not sent when statusOnly=true. This
would previously cause the action graph in Cloud to start spinning when
it shouldn't because we were only checking for statuses, and not
executing anything. To achieve this we now emit events at the task
level, not the action router level.

D) Remove old "taskComplete" et al events which are no longer used.

E) Explicitly _pick_ fields to emit, as opposed to _omitting_ from some
larger interface. The latter is not type safe and can easily cause
unwanted data to be sent "over the wire" as the "parent" interface changes
over time.