fix(cloud): emit ns statuses for in-cluster builds
We now emit a `namespaceStatus` event whenever one of our internal
helpers for checking the status of a Kubernetes namespace is called.
This is done whenever we e.g. perform a Deploy, Test or Run in that
namespace, or when we perform an in-cluster Build in that namespace
(e.g. via BuildKit or Kaniko).

Here, we also remove namespace statuses from the relevant result types
(e.g. `ServiceStatus`, `RunStatus` and `EnvironmentStatus` and the
associated schemas). These added unwanted noise to the interfaces and
schemas.

Instead, we emit the `namespaceStatus` events on the plugin event
broker (`ctx.events`) and re-emit them to `garden.events` in the
base router.