fix: do not mark deploy action outdated for irrelevant file changes
for an Deploy/Test/Run action, when build is specified, set the include to []
unless either of include or exclude is explicitly set on the config.

test: add some more unit tests for vcs tree versioning
chore: update call signature of getTreeVersion
chore: cleanup comment
chore: fix build
fix: do not mark deploy action outdated for irrelevant file changes
for an Deploy/Test/Run action, when build is specified, set the include to []
unless either of include or exclude is explicitly set on the config.

chore: fix broken tests