feat(vcs): add new git repo scanning method to improve resolution speed
This adds a new experimental git repo scanning method, enabled by setting
`scan.git.mode: repo` in the project config or by setting
`GARDEN_GIT_SCAN_MODE=repo`.

It can still do with quite a bit of optimization but the general idea is to
do far fewer (albeit generally more expensive) calls to git.

For now I don't think it makes sense to make this the default since there are
pros and cons to either approach, but for large repositories with many
actions/modules this is likely to be more efficient.

I suggest we collect user feedback and do IRL tests before diving in and
optimizing further. One obvious optimization will be to index the repository
files more efficiently, instead of just filtering on a straight list of paths.
It would also make sense to delay hashing files that aren't in the index until
they are called upon.