fix(server): ensure logs are displayed in dev command
This fixes and issue that was there all along but wasn't exposed until
we fixed a separate issue with the instance manager (#4649). Until then
the instance manager wasn't re-using instance properly and masking the
underlying problem.

Before this particular fix, we were setting the log level on the Garden
instances which doesn't work if we're re-using them. So either
logs that shouldn't be visible were, or vice versa.

This fixes that by not tampering with the log level on the Garden
instance but rather optionally overwriting it when running a given
command.