fix(cloud): properly handle dev delegation
When a command is run with an option that results in it being run inside
the `dev` command instead (e.g. `garden deploy --sync`), we now register
the initial command session with Cloud as a `dev` command with the
`--cmd` option set appropriately. This is useful for Cloud when listing
command sessions created this way.

For good measure, we include an explicit `isDevCommand` parameter with
the session registration request to Cloud to make this setup more robust
against future changes (instead of inferring whether the command is a
`dev` command from the `commandInfo` map).