perf(git): optimize git scan when exclude but no include filter is set
This can be a major optimization in certain cases.

The reason this wasn't done previously was due to a git ls-files issue
but we missed out on this path when no includes are used, in which case
the issue doesn't come up.

Usually includes are much more specific than excludes in configs, so this
approach should cover a lot of user scenarios.

It is still more optimal to use `.gardenignore` files over exclude patterns
where possible, but using excludes in configs is obviously  a common and
supported use case, and not always possible to use the ignore files (say,
if many actions/modules share a directory and need different exclusions).