feat(exec): add --target flag in exec command
introduces a flag --target in exec command. This is useful when a
Deploy action consists of multiple underlying components and user
wants to target a specific one. At the moment, this flag is only
supported for kubernetes-exec and container-exec.

When a pod has multiple containers, user can use target flag to exec
into a specific container. Additionally, if no container name is
provided, it respects the annotation kubectl.kubernetes.io/default-container
and defaults to that container.

The existing behavior is still the default: it picks the first container
in Pod.