fix(mutagen): too restrictive default permissions
Especially for in-cluster building, the file permissions mutagen
defaults to are too restrictive.

Git adheres to the system umask by default, which usually is set to 022.
That leads to file permissions 644 and directory permissions 755, which
means group and world readable. I would expect this to be the default on
any system.

Dockerfile's COPY directive by default copies files including the
permissions.

In Garden 0.12 we used rsync by default to transfer the build context for
in-cluster builds, and thus files had the correct permissions as rsync also
transferred the permissions.

The new default in 0.13, Mutagen, does not transfer the file permissions
by default, and thus when using in-cluster building together with the
COPY directive, the file permissions in the resulting container image
will differ compared to using a local Docker daemon.

This commit changes the default file permissions for mutagen to 644 and
the default directory permissions to 755. This is usually a good
default, and if users need different permissions they can use the
--chmod option in Dockerfile's COPY directive or the respective Garden
sync mode settings.