fix: mutagen default permissions are too restrictive (#4824)
* refactor: make sure that permission bitmask is always octal and valid

Before the code used decimals in validation, so it allowed invalid
octals.

Fix joi validation to allow maximum 0o777 for permission bitmask.

Also change the typescript type from number to an enumeration between
0o0 and 0o1000 (to also force maximum of 0o777 in typescript code).

This guards against accidentally using decimal 777, which is 0o1411 in
octal representation.

* fix(mutagen): too restrictive default permissions

Especially for in-cluster building, the file permissions mutagen
defaults to are too restrictive.

Git adheres to the system umask by default, which usually is set to 022.
That leads to file permissions 644 and directory permissions 755, which
means group and world readable. I would expect this to be the default on
any system.

Dockerfile's COPY directive by default copies files including the
permissions.

In Garden 0.12 we used rsync by default to transfer the build context for
in-cluster builds, and thus files had the correct permissions as rsync also
transferred the permissions.

The new default in 0.13, Mutagen, does not transfer the file permissions
by default, and thus when using in-cluster building together with the
COPY directive, the file permissions in the resulting container image
will differ compared to using a local Docker daemon.

This commit changes the default file permissions for mutagen to 644 and
the default directory permissions to 755. This is usually a good
default, and if users need different permissions they can use the
--chmod option in Dockerfile's COPY directive or the respective Garden
sync mode settings.

* improvmenent: good error messages when specifying an out of range mask

* improvement: use YAMLv2 octal notation style

* fix: allow octal rendering of default values in docs

* fix: remove wrong documentation in kubernetes sync schema

According to the docs, the keys

- spec.sync.paths[].fileMode
- spec.sync.paths[].directoryMode

can be used, but in fact only

- spec.sync.paths[].defaultFileMode
- spec.sync.paths[].defaultDirectoryMode

are working.

Otherwise I get

```
Error validating spec for Deploy type=container name=vote (vote): key "fileMode" is not allowed at path .sync.paths[0][fileMode]
```