fix(core): improve dependency logic for tests
Before this fix, changes to transitive (i.e. not direct) runtime
dependencies of a test action would not result in the test action being
re-run if the action handler of the intermediating runtime dependency
can return `ready` even when input versions change. This is the case for
`kubernetes` and `helm` Deploys (but not for `container` Deploys).

Note that this only applies when outdated components have already been
deployed in the environment/namespace.

To illustrate, let's say `T`, `D1`, `D2` and `B2` are a Test, two Deploys and a
Build (where `B2` is the build for `D2`).

If `B2` is changed in such a way that might cause `T` to fail when run
against an up-to-date version of `D1` and `D2` unless the action
handler for `D2` always returns a non-ready status when the input
version changes, then under those circumstances `T` would erroneously
show a cached and successful result.

This was fixed by defaulting to recursively including all runtime
dependencies for test tasks.

Also added a solver test case for this (and set up some test data &
helpers that should be useful for writing more tests that need to use
the semantics of the action-based task classes).