fix(jib): make native arm maven usable on ARM macs
The native aarch64 build of maven has not been used before this commit,
because the arch was wrongly specified as "aarch64" in the tool spec,
instead of the correct "arm64".

This commit also prevents similar mistakes in the future by improving
the types.

We are allowing platforms and architectures that are not technically supported right now here–
this can be adressed in a future PR if needed.