chore: remove the create module command (#4981)
BREAKING CHANGE: Removed the "garden create module" command.

This prevents new users getting confused and working with modules,
instead of actions as we prefer.

The alternative would have been to implement a "create actions" command
and deprecating the "create module" command; but we think the create
actions command can be implemented later as it's only nice to have.

Co-authored-by: Tim <tim@garden.io>