docs: add k8s Deploy action example with shared manifests
This commit adds an example that uses the `kubernetes` deploy action
and shares the same manifests between a couple of actions. This is
achieved by using Garden template strings in the manifests.

The example is a three tier web app version of the vote example
with "web", "API", and "DB" components that I think we
should consider using by default.

It's not as complex as the main vote example which tends to draw
attention away from the functionality being demonstrated, and not as
trivial as the basic "demo-project".

Three tier web apps are also a common pattern that a lot of people will
be familiar with.

And lastly, this is also the same example we use in the "How Garden
Works" video series so it will become familiar to users eventually.