fix: move back to js-yaml 3.14.1 for compatibility
The parsing of octal values changed in version 4.0 of `js-yaml`, the
YAML parsing library we use to load configs. Previously, values like
`0777` would be parsed as octal numbers, but in 4.0 they're parsed as
decimal numbers.

Since some users want to maintain backwards-compatibility with Garden
0.12, we're moving back to `js-yaml`  3.14.1 (the last 3.x release) to
get back to the old octal value parsing behavior.