improvement: more useful "create project" yaml (#4985)
While it is cool that we are re-using the descriptions from our docs,
there are some problems with this yaml code:
- The comments often do not add any valuable information. For example
  it's already clear that `name: foobar` is the name of the project.
- Another problem is that the description in the reference docs is not
  necessarily the information a user needs when getting started. This
becomes apparent in the comment above `apiVersion`, for example.

This commit simplifies the `create project` command implementation to
just write a file with hardcoded yaml. The comments are hand-crafted for
a good getting-started experience.

This commit also updates the docs that refer to the "create project"
command.

Co-authored-by: Tim <tim@garden.io>
Co-authored-by: Anna <anna@garden.io>