fix(framework): several error handling improvements (#5001)
* fix(framework): several error handling improvements

- Made sure that InternalError is only used for errors that "should not happen" and
  also to wrap unhandled errors that originate from libraries or the NodeJS standard library.

- Stop masking leaf errors (leaf means basically the root cause).
  One of the most occuring errors on Amplitude is "graph", but that's a
  wrapping error. This will give us better insights in what error types
  actually causing command failures.

- Consistently handle unexpected errors and format them equally on all
  levels.

- Format InternalErrors as crashes with an explanation
  and a call to action to report the bug on all levels

- Treat uncaught errors on the CLI level (cli.ts) as crashes, even if
  they are GardenError instances other than InternalError, because the GardenCli class should not
  throw but return an exit code.

* improvement: new CRASH.md issue template

* improvement: make GardenError recognition less error-prone

Previously we had an exported interface called GardenError together with
an exported abstract base class called GardenBaseError.

We had a function called isGardenError that checked for interface
conformance, and used it in some parts of the code base, and then we
also had parts of the code base that checked for `instanceof
GardenBaseError`.

This commit eliminates the GardenError interface to eliminate a class of
errors, where we fail to recognize the GardenError and treat it as a
crash.

* fix: tests

* fix: don't throw on custom command validation fail

* test: implement tests for command level errors

* test: add yaml output crash test

* fix: filter error details early to avoid issues later

Bugfix: Before this commit when running `garden deploy -o json`
in the Garden root project, the json was several thousands lines long

It was a little bit tricky to solve the error, because I ran into infinite recursion
issues between `toSanitizedValue` and  `sanitizeValue`

Now the -o json output looks good to me.

This commit also removes dead code in the GraphNodeError.

* test(solver): add test for wrapping behaviour

* fix: do not crash when spawned child processes fail

Usually spawned child processes are expected to fail for various
reasons.
Wrap them in ChildProcessError

Also improved the error message and error type for Kubernetes apply
through kubectl.

* refactor: remove

* fix: handle undefined message in wrapError

* fix: lint errors

* fix: avoid throwing non-garden errors

* fix: lint

* fix: error handling issues

* improvement: remove "sorry"