improvement: allow dev@garden.io to impersonate service account for tests
This allows google artifact registry access in the garden-ci project.
Using the service account impersonation has the benefit that we can allow any google group to impersonate this service account, so we do not need to manage access manually:
This only works because we granted the google group "dev@garden.io" permission to impersonate this service account with the following command:
`gcloud iam service-accounts add-iam-policy-binding gar-serviceaccount@garden-ci.iam.gserviceaccount.com --member=group:dev@garden.io --role=roles/iam.serviceAccountTokenCreator`
We could extend this in the future to allow access to the CI cluster in the same way, without the need to manage users manually.

Co-authored-by: twelvemo <anna@garden.io>