docs: update welcome page (#5117)
This new welcome page was built via the Gitbook editor and contains
cards that require HTML elements.

Due to how Gitbook is currently set up, we can't make changes to the
main docs page there because it would essentially merge the changes back
to our main branch which is protected.

So instead these changes where made in a test project and copied over.

That's obviously an awful workflow and we need to figure out a proper
solution here that allows us to edit on Gitbook without breaking branch
protection rules.

I just took the hacky route for now so that we can get these changes in.