fix(k8s): handle AEC-paused resources properly (#5122)
* fix(k8s): handle AEC-paused resources properly

This is a follow-up quickfix for #4846.
It reads a specific annotation and checks if it has a special value set by Cloud.
This needs to be changed to reply on a more generic and reliable way of k8s resource comparison.

* refactor(k8s): explicit typing

* refactor(k8s): extract local variable

To avoid repetitive creation of the same immutable string.

* chore: add TODO comment

* refactor(k8s): introduce named constant for `manifest-hash` annotation key

* refactor(k8s): use helper to construct the annotation key

* chore: helper function to check sha256 validity

* refactor: move hash-helper function to a dedicated module

* fix(k8s): smarter manifest hash checking

Consider any non-sha256 value as an outdated marker for k8s resources.
This covers the current AEC behavior on the Cloud side.
This also allows users to do manual modifications of k8s resources.

* test: add integration test

* test: fix integration test to restore the valid cluster state

* fix: re-create `Regex` object before every `test` call

Because `Regex` is a stateful object.

* test: use separate module config in integ test

To avoid unwanted modifications of the remote resources used by the other tests.

* chore: typo fix

Co-authored-by: Shumail Mohyuddin <shumailmohyuddin@gmail.com>

---------

Co-authored-by: Shumail Mohyuddin <shumailmohyuddin@gmail.com>