docs: centralize `contributor` docs and provide on-ramp (#5108)
* chore: linting and re-org

To reduce docs sprawl and provide a more discoverable on-ramp for internal and external contributors to Garden, I've broken out and centralized contributor docs under `docs`.

* docs: add section for contributors to docs

This commit removes top-level headings already enforced by `title:` and includes a complete Contributing to the Docs page intended to help guide new external and internal contributors to our documentation who may or may not be Git-adept. It is also intended to serve as supporting documentation for any technical writer role we may onboard in the future.

* Add `markdownlint` suggestion

* chore: add yarn.lock to gitignore

* docs: refresh ToC

* chore: remove check-licenses

This check was preventing my push. See
https://github.com/garden-io/garden/pull/5182#issuecomment-1748492365

* chore: fix dead links