improvement(core): log stderr from exec as info (#5227)
Before this change, we logged stderr streams from
init scripts as errors but now we log it at the info level.

The reason is that some tools, e.g. gcloud, write to stderr even if it's
not an actual error.

In the case of an actual error the script exits with a non-zero
code and that's handled specifically.

So this is really just a cosmetic change to how exec provider logs are
rendered.