fix: allow generated files in build actions (#5230)
In the container and jib-container build actions, we checked for the
existance of files checked in to version control for certain checks.

Garden allows for files to be injected into the build action dynamically
at run time, e.g. by using `copyFrom` and `generateFiles`.

We want to check for actual presence of files in the action's build
directory instead of checking for the files that are included from
version control.

Fixes #5201