improvement(api): send action type to cloud
Before this improvement we weren't sending the action type to Cloud due
to some legacy issues which have since been resolved.

The action type is used in Cloud in e.g. the graph UI and on command
results.

Note that Cloud has already been updated to handle the action type and
those changes rolled out to all users, so this is only required Core
side work.