test: use more precise time comparison
This approach covers the edge case
when 2 numeric values are in the same precision vicinity,
but get rounded to the different values.

E.g. `a = 0.514` and `b = 0.516`.
The diff is `0.002`,
but the diff between rounded values was `0.01` before this fix.

This is a tiny fix, but without it, the CI fails from time to time.
That could cost ~35m of waiting time in the merge queue.