improvement(logs): add 'rawMsg' field to logEntry
This commit adds a 'rawMsg' field to the log entry message type which
can be used for log lines that don't map well between the terminal and
the web dashboard.

The dashboard will prefer the 'rawMsg' if set.

This is e.g. used when streaming service logs but prior to this commit
we added ansi colors to them which didn't play when e.g. rendering JSON
logs in the dashboard.

The commit also use this field for rendering the command start/finish
messages which are printed in the dev console and don't render nicely in
browser when the divider is too long.

This is perhaps a bit hacky and we should probably just a have a log entry
type "separator" that the browser can handle properly. But this
does the trick for now.