fix(logger): log exec build/deploy actions at info level
Previously these actions were logged at the verbose level which doesn't
really make sense imo.

In particular since other exec actions are on the info level so to the
end user it'll be confusing that some exec results get logged and others
not.

Also updated the various places where we render a divider to make sure
we also include a browser friendly version.